+++ client/v2/autocli/builder.go
@@ -6,8 +6,6 @@ import (
 
 	"cosmossdk.io/client/v2/autocli/flag"
 	authtx "cosmossdk.io/x/auth/tx"
-
-	"github.com/cosmos/cosmos-sdk/client"
 )
 
 // Builder manages options for building CLI commands.
@@ -18,9 +16,6 @@ type Builder struct {
 	// GetClientConn specifies how CLI commands will resolve a grpc.ClientConnInterface
 	// from a given context.
 	GetClientConn func(*cobra.Command) (grpc.ClientConnInterface, error)
-
-	// ClientCtx contains the necessary information needed to execute the commands.
-	ClientCtx client.Context
 
 	// TxConfigOptions is required to support sign mode textual
 	TxConfigOpts authtx.ConfigOptions