+++ client/v2/CHANGELOG.md
@@ -41,7 +41,15 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Features
 
 * [#18461](https://github.com/cosmos/cosmos-sdk/pull/18461) Support governance proposals.
-* [#19039](https://github.com/cosmos/cosmos-sdk/pull/19039) add support for pubkey in autocli.
+* [#19039](https://github.com/cosmos/cosmos-sdk/pull/19039) Add support for pubkey in autocli.
+
+### Improvements
+
+* [#19060](https://github.com/cosmos/cosmos-sdk/pull/19060) Use clientCtx from root command in autocli commands.
+
+### Bug Fixes
+
+* [#19060](https://github.com/cosmos/cosmos-sdk/pull/19060) Simplify key flag parsing logic in flag handler.
 
 ### API Breaking Changes